fix: avoid uncaught exception when client-secret.json is missing
Instead of checking for file existence and then reading it,
the read is attempted directly and if it fails, an error is
signalled to Companion via self.log() and self.status().

These two functions are used in the OBS module. log()
displays a log message for this module with a given severity
and status() shows initialization result in the 'Status'
column of the loaded module overview.