feat: replace config files with Companion config fields
This was recommended in https://github.com/bitfocus/companion-module-youtube-live/issues/9 and
indeed, the module now feels a little cleaner and more user-friendly.

Instead of using a working directory and storing raw client-secret.json
from Google API console there, individual fields are created
in the GUI for Client ID, Client secret and redirect URL.

The same applies to bearer token from OAuth2 client, however
it is still stored as a stringified OAuth2 credentials object.

Additionally, to emulate a "login" button, a special value of
the token contents is now recognized - 'login'. This triggers
the OAuth authorization request that previously happened automatically.